initial common module with several classes moved to a new location
