+++ api/src/main/java/org/bf2/operator/resources/v1alpha1/ManagedKafkaCondition.java
@@ -16,6 +16,7 @@ public class ManagedKafkaCondition {
     public enum Type {
         Installing,
         Ready,
+        Deleted,
         Error;
     }
