+++ common/pom.xml
@@ -8,27 +8,16 @@
         <artifactId>kas-fleetshard</artifactId>
         <version>1.0.0-SNAPSHOT</version>
     </parent>
-    <artifactId>kas-fleetshard-api</artifactId>
+    <artifactId>kas-fleetshard-common</artifactId>
 
     <dependencies>
         <dependency>
-            <groupId>io.dekorate</groupId>
-            <artifactId>crd-annotations</artifactId>
-            <optional>true</optional> <!-- we need to keep this optional to avoid conflicts with quarkus-kubernetes -->
+            <groupId>org.bf2</groupId>
+            <artifactId>kas-fleetshard-api</artifactId>
         </dependency>
         <dependency>
-            <groupId>io.sundr</groupId>
-            <artifactId>builder-annotations</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.projectlombok</groupId>
-            <artifactId>lombok</artifactId>
-        </dependency>
-
-        <dependency>
-          <groupId>io.fabric8</groupId>
-          <artifactId>kubernetes-server-mock</artifactId>
-          <scope>test</scope>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-kubernetes-client</artifactId>
         </dependency>
     </dependencies>
