+++ common/src/main/java/org/bf2/common/AgentResourceClient.java
@@ -1,4 +1,4 @@
-package org.bf2.operator.clients;
+package org.bf2.common;
 
 import javax.enterprise.context.ApplicationScoped;
 
@@ -7,6 +7,8 @@ import org.bf2.operator.resources.v1alpha1.ManagedKafkaAgentList;
 
 @ApplicationScoped
 public class AgentResourceClient extends AbstractCustomResourceClient<ManagedKafkaAgent, ManagedKafkaAgentList> {
+
+    public static final String RESOURCE_NAME = "managed-agent";
 
     @Override
     protected Class<ManagedKafkaAgent> getCustomResourceClass() {
