+++ operator/src/main/java/org/bf2/operator/clients/KafkaResourceClient.java
@@ -3,6 +3,8 @@ package org.bf2.operator.clients;
 import javax.annotation.PostConstruct;
 import javax.enterprise.context.ApplicationScoped;
 
+import org.bf2.common.AbstractCustomResourceClient;
+
 import io.fabric8.kubernetes.client.dsl.MixedOperation;
 import io.fabric8.kubernetes.client.dsl.Resource;
 import io.fabric8.kubernetes.client.dsl.base.CustomResourceDefinitionContext;
