+++ sync/pom.xml
@@ -22,6 +22,10 @@
             <artifactId>kas-fleetshard-api</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.bf2</groupId>
+            <artifactId>kas-fleetshard-common</artifactId>
+        </dependency>
+        <dependency>
             <groupId>io.quarkus</groupId>
             <artifactId>quarkus-container-image-jib</artifactId>
         </dependency>
